requirements.txt
pyyaml
rich
pysubs2
ffmpeg-python
requests
pytest
